Move WriteTest in its own package, make some tests parameterized
